requirements.txt
scikit-learn
mlflow
pandas
numpy
matplotlib
seaborn
scipy
joblib
xgboost
lightgbm
catboost
tensorflow
keras
torch
genism
fastai
pytorch-lightning
optuna
hyperopt
sklearn
scikit-optimize
scikit-learn-extra
imbalanced-learn
shap
eli5
lime
pdpbox
shapash
nltk
dtreeviz
yellowbrick
mlxtend
scikit-plot
torch
torchvision
torchtext
pytorch
tensorflow
tensorflow_hub
tensorflow_datasets
keras
keras-tuner
keras-rl
keras-vis
keras-preprocessing
keras-applications
keras-embeddings
keras-ocr
keras-ocr
